split format.js into a few files
